refactor: Align auth client code.
