+++ app/root.tsx
@@ -14,7 +14,7 @@ import { type ReactNode, useEffect } from 'react';
 import { GeneralErrorBoundary } from '#components/error-boundary';
 import UI from '#components/ui';
 import { getUser } from '#utils/auth/auth.server.ts';
-import { useAuthBroadcast } from '#utils/auth/user';
+import { useAuthBroadcast } from '#utils/auth/auth.ts';
 import { combineHeaders } from '#utils/misc.ts';
 import { ClientHintsFallback } from '#utils/theme/client-hints-fallback';
 import { getHints } from '#utils/theme/client-hints.server';
